cleaning up project
