+++ src/stateMachines/Scheduler.java
@@ -1,4 +1,4 @@
-package elevators;
+package stateMachines;
 
 import java.util.LinkedList;
 import java.util.Queue;
