+++ src/tests/SchedulerTest.java
@@ -9,7 +9,7 @@ import java.util.Queue;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import elevators.Command;
+import commands.Command;
 import elevators.Elevator;
 import elevators.Scheduler;
 
@@ -54,7 +54,7 @@ class SchedulerTest {
 	 * Test adding a command to the Queue
 	 */
 	void testAddCommand() {
-		Command cmd = new Command(Calendar.getInstance(), 1, Command.Direction.UP, 123);
+		Command cmd = new Command(Calendar.getInstance(), 1, commands.Direction.UP, 123);
 		assertNotNull(cmd);
 		this.testScheduler.addCommand(cmd);
 		assertTrue(this.testScheduler.getCommandQueueSize() == 1); // should have 1 item in 
@@ -65,7 +65,7 @@ class SchedulerTest {
 	 * Test removing a command from the Queue
 	 */
 	void testRemoveCommand() {
-		Command cmd = new Command(Calendar.getInstance(), 1, Command.Direction.UP, 123);
+		Command cmd = new Command(Calendar.getInstance(), 1, commands.Direction.UP, 123);
 		this.testScheduler.addCommand(cmd);
 		cmd = this.testScheduler.getCommand();
 		assertNotNull(cmd);
